divided grawToEventTPC into a header, source, and a file with main function added readNEvents option to config puts a limit on a number of events converted by grawToEventTPC changed the do while loop into a for loop in grawToEventTPC
